requirements.txt
absl-py==2.0.0
aiofiles==23.2.1
aiohttp==3.8.6
aiosignal==1.3.1
alembic==1.12.1
anycache==2.1.0
anyio==3.7.1
app==0.0.1
appnope==0.1.3
asgiref==3.7.2
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
b==1.0.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
blinker==1.6.3
botocore==1.31.74
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.0
cheese==0.1.0
click==8.1.7
cloudpickle==3.0.0
comm==0.2.0
contourpy==1.1.1
coverage==7.3.2
crtime==0.0.6
cryptography==41.0.4
cssselect==1.2.0
cycler==0.12.1
dask==2023.10.1
decorator==5.1.1
deprecated==1.2.14
desktop==0.4.2
dill==0.3.7
django==4.2.6
executing==2.0.1
filelock==3.13.1
flask==2.3.2
flask-bcrypt==1.0.1
flask-migrate==4.0.4
flask-sqlalchemy==3.0.3
flatbuffers==23.5.26
fonttools==4.43.1
frozenlist==1.4.0
fsspec==2023.10.0
future==0.18.3
gast==0.5.4
google-auth==2.23.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
gorilla==0.4.0
greenlet==3.0.1
grpcio==1.59.2
guineapig==0.3.0
gunicorn==20.1.0
h11==0.12.0
h5py==3.10.0
html2text==2020.1.16
httpcore==0.13.7
httpx==0.18.2
idna==3.4
importlib-metadata==6.8.0
ipython==8.17.2
ipywidgets==8.1.1
itsdangerous==2.1.2
jedi==0.19.1
jinja2==3.1.2
jmespath==1.0.1
jupyterlab-widgets==3.0.9
keras==2.14.0
kiwisolver==1.4.5
libclang==16.0.6
locket==1.0.0
lxml==4.9.3
mako==1.2.4
markdown==3.5
markupsafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
ml-dtypes==0.2.0
more-itertools==10.1.0
multidict==6.0.4
mymodule==1.0.0
mysql-connector-python==8.2.0
nap==2.0.0
nptime==1.1
numpy==1.26.1
numpy-indexed==0.3.7
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.2
parso==0.8.3
partd==1.4.1
pexpect==4.8.0
pillow==10.1.0
prompt-toolkit==3.0.39
protobuf==4.21.12
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pygithub==2.1.1
pygments==2.16.1
pyjwt==2.7.0
pynacl==1.5.0
pyparsing==3.1.1
pypi-search==2.0
python-dateutil==2.8.2
pytz==2023.3.post1
pyyaml==6.0.1
rdltr==0.1.14
readability-lxml==0.8.1
recommend==0.2.2
requests
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.9
schedule==1.2.1
scipy==1.11.3
seaborn==0.13.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
sqlalchemy==1.4.47
sqlparse==0.4.4
stack-data==0.6.3
tabulate==0.9.0
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.3.0
testpy==0.3
toolz==0.12.0
tqdm==4.66.1
traitlets==5.13.0
typeguard==4.1.5
typing-extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
wcwidth==0.2.9
werkzeug==3.0.1
widgetsnbextension==4.0.9
wrapt==1.14.1
yarl==1.9.2
zipp==3.17.0
matplotlib-venn==0.11.9
psutil==5.9.8